requirements.txt
Flask==2.3.3
requests==2.31.0
stripe==7.11.0
supabase==1.2.0
python-dotenv==1.0.0
httpx==0.24.1